requirements.txt
uvicorn
fastapi
reportlab
valkey
requests
celery
celery[redis]
jinja2
python-decouple
